Clean Code 修改
# message auto-generated for no-merge-commit merge:
merge master_位运算 into master

Clean Code 修改

Created-by: msprofDeveloper

Match-id-6c0f184173b7b801bd8851f939e011c03075f395

+++ analysis/csrc/parser/chunk_generator.cpp
@@ -54,11 +54,11 @@ int ChunkGenerator::ReadChunk()
             return ANALYSIS_ERROR;
         }
         auto fileSize = File::Size(file);
-        if (fileSize % chunkSize_ != 0) {
+        if (fileSize % static_cast<uint64_t>(chunkSize_) != 0) {
             ERROR("The read chunk file size is invalid: %", file);
             return ANALYSIS_ERROR;
         }
-        remainSize_ += fileSize / chunkSize_;
+        remainSize_ += fileSize / static_cast<uint64_t>(chunkSize_);
     }
     return ANALYSIS_OK;
 }
